requirements.txt
openai>=1.30.0
boto3>=1.34.0
streamlit>=1.35.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
